+++ packages/style/storybook/stand-alone-documentation/colors/assets/index.js
@@ -1,4 +1,0 @@
-import cardColorsExample from "./cardColorsExample.png";
-import menuColorsExample from "./menuColorsExample.png";
-
-export { cardColorsExample, menuColorsExample };