+++ packages/style/storybook/stand-alone-documentation/welcome/assets/index.js
@@ -1,7 +1,0 @@
-import Orr from "./Orr.png";
-import Hadas from "./Hadas.png";
-import Meirav from "./Meirav.png";
-import Sergey from "./Sergey.png";
-import Yael from "./Yael.png";
-
-export { Hadas, Orr, Meirav, Sergey, Yael };