+++ packages/style/storybook/stand-alone-documentation/welcome/helper-components/Icons/IconContainer.tsx
@@ -1,7 +1,0 @@
-import "./IconContainer.scss";
-
-export const IconContainer = ({ icon }: { icon: React.FC }) => {
-  return <div className="monday-ui-style-storybook_icon-container">{icon({})}</div>;
-};
-
-export default IconContainer;