+++ packages/style/storybook/stand-alone-documentation/welcome/intro/intro.jsx
@@ -1,7 +1,0 @@
-import styles from "./intro.module.scss";
-
-export const WelcomeIntro = () => (
-  <span className={styles.intro}>
-    Here you can find our design system css tokens and instructions on how to use them
-  </span>
-);