+++ packages/core/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.95.0](https://github.com/mondaycom/monday-ui-react-core/compare/monday-ui-react-core@2.94.0...monday-ui-react-core@2.95.0) (2024-03-12)
+
+
+### Features
+
+* **AvatarGroupCounter:** pass noAnimation prop ([#2010](https://github.com/mondaycom/monday-ui-react-core/issues/2010)) ([048b163](https://github.com/mondaycom/monday-ui-react-core/commit/048b1639fe6f648450802889624d876c181fc12f))
+* **Table:** new sort behavior ([#1935](https://github.com/mondaycom/monday-ui-react-core/issues/1935)) ([32790f4](https://github.com/mondaycom/monday-ui-react-core/commit/32790f4309daf9bd38c9eeb62e6b11cbd52baea0))
+
+
+
+
+
+# [2.94.0](https://github.com/mondaycom/monday-ui-react-core/compare/monday-ui-react-core@2.93.0...monday-ui-react-core@2.94.0) (2024-03-05)
+
+
+### Bug Fixes
+
+* **TableHeaderCell:** change hover color ([#1888](https://github.com/mondaycom/monday-ui-react-core/issues/1888)) ([c057a2e](https://github.com/mondaycom/monday-ui-react-core/commit/c057a2e83d2d3fb62c05e29ab768f44a77045fe9))
+
+
+### Features
+
+* **Flex:** new align's value - baseline ([#1989](https://github.com/mondaycom/monday-ui-react-core/issues/1989)) ([224be04](https://github.com/mondaycom/monday-ui-react-core/commit/224be042e0265740d174f995e290f3cefa36ed1b))
+
+
+
+
+
 # [2.93.0](https://github.com/mondaycom/monday-ui-react-core/compare/monday-ui-react-core@2.92.9...monday-ui-react-core@2.93.0) (2024-02-28)